Basic Support for Frames
Only the focused frame will act on key commands.

On some sites (such as Gmail), the main frame is already focused, so
commands just work.

For other sites, focusing the desired frame may be necessary, which can
be done with the Tab key (not optimal) or by clicking with the mouse
(even less optimal).

An additional command to cycle through frames is desirable, which will
likely come in a future commit.