Command to go to next frame.
I remapped toggleViewSource to 'gs' and set 'gf' to nextFrame.

Sorry this is such a huge commit. This is really the simplest way I can
find to allow the extension to track all available frames as well as the
currently-focused frame.

If Chrome would allow access to window.frames[i], then this could
probably be simpler.