Make "d" and "u" scroll up and down a page instead of closing and restoring tabs. This implements #274.
The motivation behind this change is that scrolling a page is far more common than
closing a tab, and so those actions should be on a hotkey which doesn't require a modifier.
Once you get used to scrolling without hitting ctrl, you won't be able to go back.

For existing users, this change also adds a custom mapping for them in their preferences to preserve
the old behavior as they upgrade from 1.21 to 1.22.