Prevent bookmark completion box from overflowing.
Happens when the partial string input matches more results
than can be displayed vertically. This causes the input section
at the top, together with the top matching URLS to disappear.

Solution is to only center the box when it's height is less than
the window height, otherwise it's top is at coordinate 0.
No negative top.