[RFC] Use same hint string for the same links
This reduces the total number of hints used - helpful on a page where there are
more than one links that point the the same location.

For example reddit.com main page has many <a> links that point to same location
in both the thumbnail and the text description of a post.

Possible caveat:

May break javascript links like <a href="dosomething()"> where dosomething()
does something depending on the nth-order of the target element.

Links like these are also possible problems: <a href="#">, where there are
javascript event listeners attached to the element.

If anyone can suggest a way to correctly detect these kind of gotchas then
that'd be great as then we can ensure that those link will not be considered as
pointing to the same target, something along the example below but need to
handle other possible cases:

function isJavascriptURI(href) {
	return href.match(/^javascript:/);
}