Reimplement out-of-order link hints generation
The original implementation (325e8b8e3b31ae2a7379f32e657bdade8a2efe0d)
generates unnecessary trailing characters in the hint strings due to the
reversal of the hint strings.

A solution for removing the trailing characters using trie was suggested in
pull request #424 but was deemed too complicated for the goal.

This commits achieves similar "spread-out effect" without reversing the hint
string using a simpler method than the "complicated-fix-on-top-of-a-hack" trie one.