Refactor settings storage and make it support empty strings.
It appears that localStorage keys with the empty string as their value
will have their values changed to undefined after a browser restart. The
DOM Inspector shows that the keys are still present, but '{{key}} in
localStorage' returns false. Convert all localStorage values to JSON as
a workaround.

This allows us to store null, numerical etc values seamlessly.

Closes #434.

Disable Vimium in the options page, due to a name collision.