Fix selector specificities.
The selector in userDefinedLinkHintCss for non-matchingCharacters takes
precedence over our default selector for matchingCharacters. We do not
want this override to happen.

This problem was probably not observed previously because the
userDefinedLinkHintCss was being loaded earlier in the page rendering,
so our default selectors managed to override it when they were loaded
later.