A WIP rewrite of completion in the vomnibar.
The purpose of this refactor is to simplify the contract so it's easier
to modify, and to make some substantial usability improvements.

One of the key differences is that matching is no longer fuzzy. If you
want to search more than one term, separate them by spaces. This matches
the behavior in Firefox.

While fuzzy matching is a nice experience for a limited set of known
items (like files in the current project), it doesn't work well for
a huge messy collection, like URLs in your history. The query "hello"
will match random stuff from your google search results for instance,
and it's hard to prune that noise using ranking intelligence.