vomnibar: New tabs for 'O' and 'B'
'B' is a new command for searching for a bookmark and opens it in a new
tab.

'O' now searches through history and bookmarks and opens the selected
item in a new tab. Previously, it started vomnibar with the current URI
selected and opened the selection in the same tab.

This fixes #594.

Signed-off-by: Carl Helmertz <helmertz@gmail.com>