Make the default value for scrollStepSize a string.
This interacts with these lines from settings.coffee:

    don't store the value if it is equal to the default, so we can change the defaults in the future
    if (value == @defaults[key])
      @clear(key)

If the default value is numeric, then this test NEVER succeeds (because
"==" is compiled to "===", and the types don't match).

So, scrollStepSize is stored in localStorage even if it has its default
value.  Which obviates the intention of the line quoted above.