Don't suppress the escape event in insert mode
Facebook.com, for example, uses the ESC key to close chat tabs. Since
insert mode is active whenever a chat tab has focus, Vimium was blocking
that shortcut.  This change allows the page to process the escape event
after Vimium has used it to modify focus. If the user doesn't want to
send an ESC keydown to the page (for example, to leave an fb chat tab
but keep it open), the <C-[> shortcut is still availble.