Fix some bookmarklets ruining the vomnibar UI.
If a bookmarklet contains html in a string, then that html is rendered when the
bookmarklet is filtered from the vomnibar. Because of this, some bookmarklets
are causing a mess in the vomnibar's results UI. This fixes it by running
htmlEntities on the url before display. It doesn't affect the dispay of normal
bookmarks.

The htmlEntities should be run before the highlightTerms call, otherwise the
markup added by highlightTerms will also be escaped which is not inteded.