Reload content scripts after install/update
This loads the content scripts for pages when the extension is
installed or updated (or reloaded in developer mode). This means that
the new version of Vimium will become active immediately after loading,
and so tabs won't have to be refreshed.

As a devloper, this is amazing. Refreshing each tab individually after
each code change was a pain.