Skip over tiny frames when calling focusThisFrame
GMail and Google Inbox both use a large number of small (typically 0x0
or 1x1 frames) which we have to pass over with nextFrame to get to the
frames we want. We avoid this issue by ignoring all frames with height
or width < 3, since these will be essentially invisible to the user
anyway.