Fix vomnibar flicker after tabs mode.
In tabs mode, the vomnibar is pre-populated when the query is empty.
If, as part of hiding the vomnibar, we reset it, it becomes populated
again, so the display style is reset from "none" to "block".  Therefore,
the completion list is briefly visible when the vomnibar is later
reactivated.

Solution:
- Do not run `@update()` from `@reset()`.