Modes; handle normal mode return values.
Up until this point in the development of modes, we've just let the
normal mode handlers return whatever they previously would have
returned.  This allowed keyboard events to continue bubbling down the
stack, but didn't matter, because normal mode is the last keyboard
handler on the stack.

This changes that.  Now, normal-mode key handlers return the right value
to have the handler stack stop or continue bubbling, as appropriate.