Minor refactoring of isSelected().
There's no point in focusing element if it's already active.
Also, we leave the caret in place, not just a range selection.

isSelected makes the wrong decision if the user has placed the caret at
the start of element.  We cannot distinguish that case from the user
having made no selection.  Nevertheless, this is substantially better
than the existing UX, which *always* moves the caret to the end.