Fix scrolling issue.
On this page:
  - http://www.steephill.tv/dubai-tour/

I was seeing a scrolling problem in my live browser (j/k didn't work
until I clicked on the page), but couldn't reproduce it directly in my
test browser.  That's worrying.

Nevertheless, what was happening was that we succeeded in scrolling a
test amount in one direction, but the scroll to revert failed.  The
consequence was that we concluded (incorrectly) that the element doesn't
scroll.  And j/k scrolling is broken as a result.

How can it be that a scroll in one direction succeeds, but the reverse
does not?

This fixes the problem by not checking whether we are able to undo the
scroll of our test amount.