Revert "Fix scrolling issue."
This reverts commit 90d2addc9b8f95f9272f8c2a77bdaf9dfebc0fa8.

Nope.  This is bad.  It fixes the issue referred to in the commit
record, but breaks scrolling out of text areas.

With 90d2addc9b8f95f9272f8c2a77bdaf9dfebc0fa8:

- Start editing in a text area, add enough text that it scrolls.
- `Escape`
- `k`, `k`, `k`, ...

We expect first the text areas to scroll, then -- when it reaches the
top -- the document to scroll.  However, with
90d2addc9b8f95f9272f8c2a77bdaf9dfebc0fa8, we get stuck in the text
area.