Visual/edit modes: always fully remove the selection on exit from visual mode.
There's considerable discussion in #1441 as to what we should do with
the selection on leaving visual mode.  Good arguments have been made as
to why we should keep the selection.  However, at this point,
keep-it-simple seems like the best strategy, and wholly removing the
selection (probably) provides fewer ways for the user to shoot
themselves in the foot.