Fix search-engine auto selection.
In #1389/c52c0ea57f86c1c5a132819fe85e763db84ce712 we added descriptions
to search engines. We (I) omitted to realise that the "type" of a
selection (in particular, the constant "search") is also used to
determine whether the suggestion is automatically selected by the
vomnibar.

This fixes that, such that custom search engines with descriptions are
now automatically selected in the vomnibar.