Fix error on yank/scroll.
When we exit visual mode with "y"...

"y" is a command for visual mode.  Currently it clears the selection.
Because it's executed as a command, scrollIntoView is called after
running the command.  Because the selection is cleared, scrollIntoView
found nothing to scroll, and was creating an error.

So we need to check whether there's anything to scroll into view, before
trying to scroll it.

Also fix mis-named file in the tests.