Fix(?) issue with browser icon.
I'm seeing an intermitent issue whereby the browser icon is showing as
disabled, when vimium is in fact enabled.  It's intermitent, so
difficult to track down.

I went back to look at the old code for setting the browser icon, and
reset this code to look more like the old code.  I suspect the problem
may be that we disabling the icon because we're getting no response from
one frame (or something like that).