Deactivate global insert mode on focus.
Rationale.  There are really two insert modes: the one we activate with
"i" and the one that's activated when a focusable element receives the
focus.  This makes that distinction clearer.  And, in particular, it
means that, when the latter type of insert mode is active, we *know*
that gloabl insert mode is *not* active.