Find (from visual mode) returns to viewport on no match.
The intention here is that, under visual mode, a find query with no
matches leaves us at some aribtrary place in the page.  Whereas the user
was probably looking at the text they're interested in in the viewport
to begin with.  This PR returns to the original viewport in such cases.

(In this commit, return-to-viewport is enabled for *all* finds, not just
under visual mode.  I want to try it out like this to see what it feels
like.  This is probably not the right UX.  It is not what Chrome does.)