Do not include duplicated texts in link hints.
This only effects link hints with "Use the link's name and numbers for
link-hint filtering" enabled.

We have been matching the *entire text content* of each link-hint
element.

With two (or more) hints, and with one of the elements a descendent of
the other, we have been using the entire text content of the outer node
(which includes the text content of the inner node).  This leads to odd
situations where the inner element cannot be selected just by typing its
text, because its text is a substring of the outer element's text.

For example, on Google calendar, the "Today" button shows up as two
hints, one inside the other.  Typing "today" never disambiguates the
hint.  You always have to hit enter.

There's another nasty example on feedly, where an outer container is
clickable, but its text contains all of the (many) texts of the (many)
contained links.  So the hint always has to be selected manually.

Here, when generating the text for an element, we exclude the texts from
any descendent node which has already been considered.