Reload content scripts after install/update.
This is @mrmr1993's work from #1041.

Reload content scripts when vimium is installed or updates.

(@mrmr1993:  The automatic merge was really messy (or, at least, I
couldn't figure out what was going on).  Since the bulk of #1041 was
actually quite compact, I took the liberty of just copying it in.  Hope
you don't mind.)