Propagate exclusion rules to all frames.
When the active tab changes, we call updateActiveState to update the
icon and propagate any changed exclusion-rule state to the tab.  All
frames received the request.  However, only one response is received by
the background page.  Therefore, new exclusion rules are only propagated
to one frame.

Here's what can go wrong...

On gmail, open the hangouts frame.  Add an exclusion rule sepcific to
the hangouts frame.  Save it.  The update is propagated only to the main
frame.  The new exclusion rule is not in effect in the hangouts frame.
That's wierd and obviously wrong.

In this commit, every frame receiving the getActiveState request also
calls checkIfEnabledForUrl to ensure that any new exclusion-rule state
is propagated.  This is overkill, to some extent.  We should really move
settings to chrome.storage and have each frame check locally for changes
affecting it.