Activate vomnibar in window.top.
This changes vomnibar commands to activate not in the current frame, but
in window.top.  Consequently, the vomnibar always appears in the same
position, and we don't get odd looking vomnibars in small frames.

Apart from the better UX, this seems to be the right thing to do.
Vomnibar commands apply to tabs (not frames).

Currently incomplete:

- On exit, the focus is not returned to the frame which originally
  had the focus.  (It's returned to window.top).

- The vomnibar can lose the focus and not hide itself for various
  frame/click combinations.