Upgrade message; show after permission accepted.
We're asking for the "notifications" permission.  I'm not sure what
happens.  If the user is asked to accept the permission, then we
wouldn't show the notification until the next time Vimium starts.

Instead, we check more frequently.  So the notification will show the
next time a tab is created (or changes page).