Check isEnabledForURL on hash change.
If an exclusion rule depends on the hash/anchor, then we're not picking
it up.  Here's a concrete example of this:

   - https?://ca*.computing.dcu.ie/[0-9]*-*#*

(which matches slides prepared via "slidy").  The initial URL does not
include the anchor/hash, so we miss the exclusion rule.  The page is
bounced immediately to an anchor/hash for which the rule should apply,
and we miss it.

(There may be other ways in which the URL can change (WebNavigation?),
we need to look into this.)