Selection handling on exiting visual mode.
Previously, we removed the selection on exiting visual mode.  This
collapses the selection instead.

My own experience with visual mode suggests that this is better.  In
particular, if we remove the selection, then when the user re-enters
visual mode, they're dropped into caret mode. It is unlikely that the
default caret position selected by caret mode is better than the last
text selected by the user in either caret mode or visual mode.