Track tab URL even if frame's URL changes.
We track the URL of the active frame in each tab so that we can
correctly populate the rules in the page popup.

Previously, we tracked URL changes only on focus events.  This commit
moves that to isEnabledForURL, which is also called when the URL changes
as a result of web navigation.