Fix HUD on options page (temporary fix).
Following on from #1652, vimium was failing to initialize correctly on
the options page.

Specifically, installListeners() was running before
initializeOnDomReady(), hence before HUD.init().  installListeners()
runs new GrabBackFocus(), which hides the HUD, which is not yet
initialized - so it fails.

This is a temporary fix.  Could you take a look, @mrmr1993?  Perhaps
there is a race condition which only arises on the options page.  It may
be enough to just have HUD.hide() silently pass if @tween is not yet
set.