In vomnibar, <Enter> excludes prompted text.
This changes the behaviour of <Enter> in the vomnibar when a prompted
text is present (i.e. text from a suggestion which the user has not yet
typed.  Now, that text is no longer included in the query launched with
<Enter>.  After a little experience, I'd say that's a pretty bad UX.

This also adds a <Ctrl-Enter> binding to launch the query with the
prompted text included.