Rework some aspects of search engine completion.
- We arrange that:
   - for custom search engines, or
   - for other searches where there is at least one match
  that match is at the top of the list (relevancy 1).

This top suggestion will the suggestion which will be used to popuplated
the promted text in the vomnibar.  If we arrange that it's at the top of
the list, then it is also the text that we get in the vomnibar if we hit
<Tab>.

This leaves us in the position where:

   - If the user wants just the text they've typed, then they hit
     <Enter>.

   - If they want all of the text in the input (including the prompted
     text), then they hit <Tab>, <Enter>.  This is a very natural UX.
     It feels like <Tab> is filling in the prompted text, although all
     it's really doing is choosing a suggestion (the first one), as it
     always has done.

There is one small catch. We need to avoid a clash with the domain
completer, which also likes to force its suggestion to the top of the
list.  There's a test for that case.  Basically, we can apply this trick
if it's a custom search, or if the user has finished typing the first
query term.