Search completion; implement <Enter> timer.
With suggested completions appearing in the vomnibar, the meaning of
<Enter> can change when the completion list is updated, which happens
asynchronously.  The user can end up selecting something they didn't
intend to ... which is bad.

This implements a temporary fix.  When the completion list updates (and
the update might affect what the user expects <Enter> to do), we block
<Enter> for a short time (250ms).  That's pretty bad too.

These UX issues around search completion are proving difficult to get
right.  I will open an issue to discuss them soon.