Give the vomnibar a (kind of) history (autoSelect).
If we insert the text of hostory completions into the vomnibar input,
then when the completions change, we'll pick up the arbitrary text of
whatever happens to be the current selection.  So we need to always
reset the selection when the vomnibar updates.  This eliminates the need
for the previous autoSelect logic.