Search completion; do not adapt completion scoring.
The relevancy score of completion suggestions should not change as the
user continues to type.  See "UX Issue 1" in #1651.

(The effect of the previous implementation was that search completions
would jump, unpredictably, up the suggestion list, often displacing
suggestions from other completers for no reason which was apparent to
the user.)