Rework parsing of custom key mappings.
Most of this is just a tidy up of code that's been around for a long
time.

The only difference, however, is that now a key mapping can include
extra data ("extras") after the name of the command.  For example,

map s vomnibar.activate keyword=g

Here, and extra property "extras" is added to the command registry:

   extras: ["keyword=g"]

This is a first step towards direct vomnibar activation for custom
search.