Automatically generate documentation for completion engines.
This automatically generates a markdown document from custom-search
completion engines.

The markdown document *should* be committed to the repo.  That way, we
can create links to it from the wiki.  Ideally, we sould create a new
tag for each release, and link to that version of the markdown file from
the README.