Add a default value for helpDialog_showAdvancedCommands
We need this because Settings rejects key/value pairs from
chrome.storage for which there were no default values.

Previously, this only meant that the setting would not sync; now it
meant that the setting wasn't ever made available to the frontend. This
commit fixes it, and now the setting will sync.