Tweak vomnibar custom search behaviour.
With this, if we're in custom search mode and the cursor is at the start
of the vomnibar input and the user types <Delete>, then we exit custom
search mode and reinstate the keyword (and arrange for the cursor to be
in the right place).

This makes it possible to change (or just disable) the custom search
engine without having to first wholly delete the query.