Only offer suggestions from the current custom search engine.
When the user changes custom search engine during a single vomnibar
activation, we should only offer suggestions which match the current
search engine (not the previous one).

We now do this not just by checking that the suggestion is a custom
search suggestion, but by checking that the actual search URL matches.