Global marks; only handle messages in main frame.
On sites with several frames (e.g. Facebook), if we allow all of the
frames to handle the setScrollPosition and showHudForDuration messages,
then the focus ends up in an arbitrary frame.  And, on Facebook, say,
Vimium ends up broken until the focus is returned to the main frame.
So, we only handle these messages in the main frame.