Re-work completions: initial refactor.
The original completion-engine interface was based on three functions.
With some experience, it seems there is a pattern involving explicit
regular expressions which is used by all actual engine implementations.
This is a refactoring to make those regular expressions explicit (and
required), and is a first step towards adding additional fucntionality.

This also simplifies the completion cache key (use JSON instead of some
weird hash).