Re-work completions: extend engine wrapper to handle prefixes.
This commit contains the bulk og the material changes for which the
previous commits established the basis.

1) Add a general framework for detecting query prefixes in search URLs,
adding them to query sent to the completion engine, and stripping them
from the resulting suggestions.  This allows the user to have a
search engine...

   j: http://www.google.com/search?q=javascript+%s Javascript

and have the prefix "javascript" included (automatically) in queries
sent to completion engines, which results in substantially better
suggestions.

2) Re-work completion for Google Maps in a simpler form.