Re-work completions: only offer actual search URL matches.
When filter suggestions from other completers, most notably the history
completer, we only keep suggestions which match the current searchUrl
and completer.  Here, we also *replace* the URL of the suggestion.  With
duplicate elimination, multiple history entries (e.g. those generated
with various "Search Tools" settings on Google) will be collapsed to
one.  This matters because, with custom search engines, we don't show
the URL, so the user can't see differences in the URL.  Without this,
the user can be presented with a list of apparently identical
completions.