Speed up vomnibar load.
With #1697, the vomnibar used with a custom-search keyword is not
initially empty (it shows the primary suggestion for the custom search
engine, even if the query is empty).  The way things were structured
previously, the user nevertheless had to wait until the history cache
(which was not actually required) had been fetched before the vomnibar
was updated.

This commit just flips things around a bit such that,
   - onComplete() is called immediately if the history is not actually
     required, and
   - the history cache is primed before it is needed, so it will
     (hopefully) be available before the user's next keystroke.

This avoids a noticable delay, particularly on start up and with a large
history.

(This is part of a sequence of vomnibar UX tweaks.)