Quick-fix: check that all settings have default values.
On the options page, the outputs a console error message if there is an options-page settings which does not have a default settings value.

The idea is to prevent issues like #1731.

It would be better to have a test.  So this is really just a quick fix.  However, it does verify that there are no more missing default values.