Initialise modes and install listeners at the same time.
Previously, we initialised modes early, but then installed our listeners
only after DOM ready, and then only after we'd heard back from the
background page in `checkIfEnabledForUrl`.  This creates a gap during
which modes are installed, but they're not receiving events.

If an input is focused during that gap, then we don't pick it up when
the modes are installed, and we don't pick it up when the input is
focussed (because we're not listening).  This commit moves these two
initialisation steps together, so they happen at the same time and there
is no gap.

Fixes #1738.