Make global marks consistent with other tab creation operations.
This makes global marks use the same code for opening new tabs as other
operations.  Thus, the new tab appears in the same position as it would
if it were created, for example, via the vomnibar.

To do this properly, we move the three tab-creating functions into an
object which can be (and is) exported (TabOperations).

Also fixes these three operations such that they all take the same
arguments; that is (request, callback).  Not all of these callbacks are
being used.  But we should be consistent.