Inline final part of the background keyPort handler in pushKeyToKeyQueue
This commit causes a test to fail, but only by introducing the failing
code into the synchronous testing environment.

The failure existed previously (although as a race-condition; now it's
guaranteed), and tests were passing simply because they do not cover any
asynchronous cases.

The simple fix is to useKeydownEvents for PassKeysMode like we already
do for NormalMode.