Only apply passkeys to single character keys explicitly
Before this change, a string in KeyboardUtils.keyNames could have the
key it represents disabled if
* its letters were in alphabetical order, and
* the user has set all of its letters as passkeys, with none in between.

For example, imagining that "del" (for delete) was in
KeyboardUtils.keyNames, the passkeys "dpyl0e" would cause the delete key
to be a passkey too.

This commit fixes the issue by only applying passkeys when keyChar is a
single character.