Bound count for H and L by history length.
Bound the count for H and L by history.length - 1.  Going further back
or forward than this is always silently discarded.

This means that, in certain circumstances, you can say 1000H to go back to the first history page and 1000L to go forward to the end of the history.

(However, this doesn't work unless you're starting at the begining or end of the history.  The history API doesn't expose where in the history we currently are.)

Fixes #1691, kind of.