Show as many short hint markers as possible
The previous formula was sub-optimal; this updates the formula to get
the maximum number of short hints that we can.

Brief explaination:

Let linkCount be the number of links,
    characterCount be the number of characters.

Let digitsNeeded be such that
  characterCount^digitsNeeded >= linkCount > characterCount^(digitsNeeded - 1),
  as in the code.

Let longLen = characterCount^(digitsNeeded - 1),
    shortLen = characterCount^(digitsNeeded - 2),
  such that they represent the number of hints a letter can prefix,
  depending on whether they are used for short or long hint strings.

Then the maximum number of links we can show hints for, when shortCount
is the number of short hint prefixes, is

maxCount = shortCount * shortLen + (hintCharacters - shortCount) * longLen
         = shortCount * (shortLen - longLen) + hintCharacters^digitsNeeded

We want that linkCount <= maxCount, so

linkCount <= shortCount * (shortLen - longLen) + hintCharacters^digitsNeeded

shortCount * (longLen - shortLen) <= hintCharacters^digitsNeeded - linkCount
shortCount * shortLen * (characterCount - 1) <= hintCharacters^digitsNeeded - linkCount
shortCount * shortLen <= (hintCharacters^digitsNeeded - linkCount) / (characterCount - 1)

Since (shortCount * shortLen) is the largest number of short hints we
can show, we get that shortHintCount can be as in the code.