Ensure hint handler is always removed.
For some (unknown) reason, we do not receive the keyup event for Shift
when activating link hints.  Consequently, we are not correctly removing
the keyup handler.  And that handler is blocking subsequent keyup
events, which has the effect of endless scrolling (because the scroller
relies on keyup events to stop scrolling).

Here, we correcttly remove the keyup handler when link-hints mode exits.

Fixes #1911.