Disable position-at-end in text areas.
When we `simulateSelect` an input and the selection is at the start, we
move it to the end.  This works well for single-line inputs.  However,
the UX is *bad* for multiline inputs (such as text areas), and doubly so
if the end of the input happens to be out of the viewport.

This commit simply disables the repositioning of the selection within
text areas.