Fix visual indicator for `focusInput()`.
Currently, if an input is only partially in the view port, then the page
may scroll when it is focused and the overlays for `focusInput()` are
wonky.  See #1257.

Here, we draw the overlay around the *entire* input, instead of just
around the visible part.  Being partially visible therefore is no longer
relevant.

Fixes #1257.
Closes #1258.