BgUtils; move tabRecency to bg_tiles.coffee.
Certain background-page utilities are actually shared, and are therefore
best placed in place that reflects that.

Here, tabRecency is moved to the new gb_utils.coffee in preparation for
implementing a go-to-previous-tab command.  In particular, it is no
longer appropriate that tabRecency be embedded within the completion
code.

logMessage() from main.coffee is also a candidate for moving to
bg_utils.coffee.