Alternative method for detecting script context.
We sometimes need to know whether we're running in an extension script
and/or on the background page.

The previous method detects this by comparing the page's URL to the
extension's URL (dynamically).  In addition to being (a little) more
work, this depends upon the specific way which Chromium formats URLs.

In this commit, we instead explicitly set global variables to indicate
the script context.