Remove link-hints with queue.
Since link-hints modes now accept a count, we don't get any extra
functionality from the "with-queue" mode.  We get the same effect by
passing a rediculously large count.

This, therefore, removes a small amount of logic associated with the
with-queue mode.

On the downside, we lose the associated HUD indicator text of the "with
queue" mode.